third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/edit-dom-actions-3-expected.txt

Tests that user can mutate DOM by means of elements panel.


Running: testSetUp

Running: testEditCommentAsHTML
==== before ====
- <div id="testEditCommentAsHTML">
      <!-- Comment -->
  </div>
<!-- Comment -->
==== after ====
- <div id="testEditCommentAsHTML">
      <div foo="bar-comment">Element</div>
  </div>

Running: testEditAsHTML
==== before ====
- <div id="testEditAsHTML">
    - <div id="node-to-edit-as-html">
          <span id="span">Text</span>
      </div>
  </div>
<div id="node-to-edit-as-html"><span id="span">Text</span></div>
==== after ====
- <div id="testEditAsHTML">
    - <span foo="bar">
          <span id="inner-span">Span contents</span>
      </span>
  </div>
